Add -statelt e and -statelset e
These options make it possible to get, in a stat file, the element and
elset orientations, as taken from the input tessellation.

+++ src/neut/neut_mesh/neut_mesh_gen/neut_mesh_gen.h
@@ -110,7 +110,7 @@ extern void neut_mesh_face_boundnodecoos (struct NODES Nodes, struct MESH
 				int *showelt1d, int *showelt2d,
 				int *showelt3d, double cl, char *entity,
 				int id, char *var, double *val,
-                                double *pvalqty, char **ptype);
+                                int *pvalqty, char **ptype);
 
   extern void neut_mesh_entity_expr_val (struct NODES Nodes,
 					 struct MESH Mesh0D,
